pyproject.toml
[project]
name = "forecast-global-electricity-price"
version = "0.1.0"
description = "Analysis and forecasting of global electricity prices using energy consumption data"
authors = [
    {name = "Gaspare"}
]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.8,<=3.12.12"
dependencies = [
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "plotly>=5.0.0",
    "kagglehub>=0.1.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "jupyter>=1.0.0",
    "ipython>=8.0.0",
]

[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["src"]
